refactor: BeachDto 삭제
