+++ src/main/java/beachcombing/backend/domain/beach/controller/BeachController.java
@@ -2,7 +2,6 @@ package beachcombing.backend.domain.beach.controller;
 
 import beachcombing.backend.domain.beach.service.BeachService;
 import beachcombing.backend.domain.beach.dto.BeachFineOneResponse;
-import beachcombing.backend.domain.beach.dto.RecordByBeachFindAllResponse;
 import beachcombing.backend.global.security.auth.PrincipalDetails;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -10,7 +9,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
