--- src/main/java/beachcombing/backend/domain/beach/dto/BeachDto.java
@@ -1,12 +1,0 @@
-package beachcombing.backend.domain.beach.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class BeachDto {
-    public String name;
-}
