+++ src/main/java/beachcombing/backend/domain/beach/mapper/BeachMapper.java
@@ -1,7 +1,6 @@
 package beachcombing.backend.domain.beach.mapper;
 
 import beachcombing.backend.domain.beach.domain.Beach;
-import beachcombing.backend.domain.beach.dto.BeachDto;
 import beachcombing.backend.domain.beach.dto.BeachFineOneResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -9,11 +8,6 @@ import org.springframework.stereotype.Component;
 @Component
 @RequiredArgsConstructor
 public class BeachMapper {
-    public BeachDto toBeachDto(Beach beach){
-        return BeachDto.builder()
-                .name(beach.getName())
-                .build();
-    }
 
     public BeachFineOneResponse toBeachFindOneResponse(Beach beach){
         return BeachFineOneResponse.builder()
