+++ src/main/java/beachcombing/backend/domain/record/service/RecordService.java
@@ -1,7 +1,7 @@
 package beachcombing.backend.domain.record.service;
 
 import beachcombing.backend.domain.beach.domain.Beach;
-import beachcombing.backend.domain.beach.dto.RecordByBeachFindAllResponse;
+import beachcombing.backend.domain.record.dto.RecordByBeachFindAllResponse;
 import beachcombing.backend.domain.beach.repository.BeachRepository;
 import beachcombing.backend.domain.member.domain.Member;
 import beachcombing.backend.domain.member.repository.MemberRepository;
